Create a directory for the format validators
Otherwise it appears it is a use case - it's not, it returns an polymorphic object with behaviours.
