--- lib/postcode_validation/use_case/format_validators/be_postcode_validator.rb
@@ -1,4 +1,0 @@
-require_relative 'regex_validator'
-class BEPostcodeValidator < RegexValidator
-  REGEX = /^[1-9]{1}[0-9]{3}$/
-end
