--- lib/postcode_validation/use_case/format_validators/fr_postcode_validator.rb
@@ -1,4 +1,0 @@
-require_relative 'regex_validator'
-class FRPostcodeValidator < RegexValidator
-  REGEX = /^(F-)?((2[A|B])|[0-9]{2})[0-9]{3}$/
-end
