--- lib/postcode_validation/use_case/format_validators/gb_postcode_validator.rb
@@ -1,4 +1,0 @@
-require_relative 'regex_validator'
-class GBPostcodeValidator < RegexValidator
-  REGEX = /(?i)([A-Z]{1,2}[0-9]{1,2}[A-Z]?)\s*([0-9][A-Z]{2})/
-end
