--- lib/postcode_validation/use_case/format_validators/nl_postcode_validator.rb
@@ -1,4 +1,0 @@
-require_relative 'regex_validator'
-class NLPostcodeValidator < RegexValidator
-  REGEX = /(NL-)?(\d{4})\s*([A-Z]{2})/
-end
