--- lib/postcode_validation/use_case/format_validators/regex_validator.rb
@@ -1,6 +1,0 @@
-class RegexValidator
-  def valid? postcode
-    validation = self.class::REGEX =~ postcode
-    !validation.nil?
-  end
-end
