+++ lib/postcode_validation/use_case/validate_address/format_validator.rb
@@ -1,0 +1,25 @@
+require_relative 'format_validators/nl_postcode_validator'
+require_relative 'format_validators/be_postcode_validator'
+require_relative 'format_validators/gb_postcode_validator'
+require_relative 'format_validators/fr_postcode_validator'
+
+module PostcodeValidation
+  module UseCase
+    class ValidateAddress
+      class FormatValidator
+        def for country
+          case country
+            when 'NL'
+              FormatValidators::NLPostcodeValidator.new
+            when 'GB'
+              FormatValidators::GBPostcodeValidator.new
+            when 'BE'
+              FormatValidators::BEPostcodeValidator.new
+            when 'FR'
+              FormatValidators::FRPostcodeValidator.new
+          end
+        end
+      end
+    end
+  end
+end
