+++ lib/postcode_validation/use_case/validate_address/format_validators/be_postcode_validator.rb
@@ -1,0 +1,13 @@
+require_relative 'regex_validator'
+
+module PostcodeValidation
+  module UseCase
+    class ValidateAddress
+      module FormatValidators
+        class BEPostcodeValidator < RegexValidator
+          REGEX = /^[1-9]{1}[0-9]{3}$/
+        end
+      end
+    end
+  end
+end
