+++ lib/postcode_validation/use_case/validate_address/format_validators/fr_postcode_validator.rb
@@ -1,0 +1,13 @@
+require_relative 'regex_validator'
+
+module PostcodeValidation
+  module UseCase
+    class ValidateAddress
+      module FormatValidators
+        class FRPostcodeValidator < RegexValidator
+          REGEX = /^(F-)?((2[A|B])|[0-9]{2})[0-9]{3}$/
+        end
+      end
+    end
+  end
+end
