+++ lib/postcode_validation/use_case/validate_address/format_validators/nl_postcode_validator.rb
@@ -1,0 +1,13 @@
+require_relative 'regex_validator'
+
+module PostcodeValidation
+  module UseCase
+    class ValidateAddress
+      module FormatValidators
+        class NLPostcodeValidator < RegexValidator
+          REGEX = /(NL-)?(\d{4})\s*([A-Z]{2})/
+        end
+      end
+    end
+  end
+end
