+++ lib/postcode_validation/use_case/validate_address/format_validators/regex_validator.rb
@@ -1,0 +1,14 @@
+module PostcodeValidation
+  module UseCase
+    class ValidateAddress
+      module FormatValidators
+        class RegexValidator
+          def valid? postcode
+            validation = self.class::REGEX =~ postcode
+            !validation.nil?
+          end
+        end
+      end
+    end
+  end
+end
